Switch number type to `u32`
`f64` is problematic, since it doesn't traits that I need.

Eventually, this shouldn't be a problem here, since `f64` is the wrong
type to represent a floating point _literal_ anyway. But I don't want to
come up with a future-proof design right now, and `u32` will do for a
while.

+++ prototype/src/language/code/code.rs
@@ -44,13 +44,13 @@ pub enum Fragment {
 #[derive(Clone, Debug, PartialEq)]
 pub enum Expression {
     FunctionCall { target: usize },
-    LiteralValue { value: f64 },
+    LiteralValue { value: u32 },
 }
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum Token {
     Identifier { name: String },
-    LiteralNumber { value: f64 },
+    LiteralNumber { value: u32 },
 }
 
 #[derive(Clone, Copy, Debug, PartialEq)]
